Made changes to Formatter + TextUI
