+++ src/seedu/addressbook/Main.java
@@ -13,7 +13,6 @@ import seedu.addressbook.parser.Parser;
 import seedu.addressbook.storage.StorageFile;
 import seedu.addressbook.storage.StorageFile.InvalidStorageFilePathException;
 import seedu.addressbook.storage.StorageFile.StorageOperationException;
-import seedu.addressbook.ui.Formatter;
 import seedu.addressbook.ui.TextUi;
 
 
@@ -27,7 +26,6 @@ public class Main {
     public static final String VERSION = "AddressBook Level 2 - Version 1.0";
 
     private TextUi ui;
-    private Formatter formatter;
     private StorageFile storage;
     private AddressBook addressBook;
 
@@ -55,13 +53,12 @@ public class Main {
     private void start(String[] launchArgs) {
         try {
             this.ui = new TextUi();
-            this.formatter = new Formatter();
             this.storage = initializeStorage(launchArgs);
             this.addressBook = storage.load();
-            formatter.showWelcomeMessage(VERSION, storage.getPath());
+            ui.showWelcomeMessage(VERSION, storage.getPath());
 
         } catch (InvalidStorageFilePathException | StorageOperationException e) {
-            formatter.showInitFailedMessage();
+            ui.showInitFailedMessage();
             /*
              * ==============NOTE TO STUDENTS=========================================================================
              * We are throwing a RuntimeException which is an 'unchecked' exception. Unchecked exceptions do not need
@@ -77,7 +74,7 @@ public class Main {
 
     /** Prints the Goodbye message and exits. */
     private void exit() {
-        formatter.showGoodbyeMessage();
+        ui.showGoodbyeMessage();
         System.exit(0);
     }
 
@@ -89,7 +86,7 @@ public class Main {
             command = new Parser().parseCommand(userCommandText);
             CommandResult result = executeCommand(command);
             recordResult(result);
-            formatter.showResultToUser(result);
+            ui.showResultToUser(result);
 
         } while (!ExitCommand.isExit(command));
     }
@@ -115,7 +112,7 @@ public class Main {
             storage.save(addressBook);
             return result;
         } catch (Exception e) {
-            formatter.showToUser(e.getMessage());
+            ui.showToUser(e.getMessage());
             throw new RuntimeException(e);
         }
     }
