+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,20 +1,13 @@
 package seedu.addressbook.ui;
 
 import seedu.addressbook.commands.CommandResult;
-import seedu.addressbook.common.Messages;
 import seedu.addressbook.data.person.ReadOnlyPerson;
 
-import java.io.InputStream;
-import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.Scanner;
-
-import static seedu.addressbook.common.Messages.*;
 
 public class Formatter {
-    private final PrintStream out;
 
     /** A decorative prefix added to the beginning of lines printed by AddressBook */
     public static final String LINE_PREFIX = "|| ";
@@ -28,70 +21,29 @@ public class Formatter {
     /** Format of a comment input line. Comment lines are silently consumed when reading user input. */
     public static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
 
-    public Formatter() {
-        this(System.out);
-    }
+    private Formatter() {}
 
-    public Formatter(PrintStream out) {
-        this.out = out;
-    }
-
-
-    public void showWelcomeMessage(String version, String storageFilePath) {
-        String storageFileInfo = String.format(MESSAGE_USING_STORAGE_FILE, storageFilePath);
-        showToUser(
-                DIVIDER,
-                DIVIDER,
-                MESSAGE_WELCOME,
-                version,
-                MESSAGE_PROGRAM_LAUNCH_ARGS_USAGE,
-                storageFileInfo,
-                DIVIDER);
-    }
-    public void showGoodbyeMessage() {
-        showToUser(MESSAGE_GOODBYE, DIVIDER, DIVIDER);
-    }
-    public void showInitFailedMessage() {
-        showToUser(MESSAGE_INIT_FAILED, DIVIDER, DIVIDER);
-    }
-    /** Shows message(s) to the user */
-    public void showToUser(String... message) {
-        for (String m : message) {
-            out.println(LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX));
-        }
-    }
-    /**
-     * Shows the result of a command execution to the user. Includes additional formatting to demarcate different
-     * command execution segments.
-     */
-    public void showResultToUser(CommandResult result) {
-        final Optional<List<? extends ReadOnlyPerson>> resultPersons = result.getRelevantPersons();
-        if (resultPersons.isPresent()) {
-            showPersonListView(resultPersons.get());
-        }
-        showToUser(result.feedbackToUser, DIVIDER);
-    }
     /**
      * Shows a list of persons to the user, formatted as an indexed list.
      * Private contact details are hidden.
      */
-    private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
+    public static String formatPersonListView(List<? extends ReadOnlyPerson> persons) {
         final List<String> formattedPersons = new ArrayList<>();
         for (ReadOnlyPerson person : persons) {
             formattedPersons.add(person.getAsTextHidePrivate());
         }
-        showToUserAsIndexedList(formattedPersons);
+        return formatIndexedList(formattedPersons);
     }
     /** Shows a list of strings to the user, formatted as an indexed list. */
-    private void showToUserAsIndexedList(List<String> list) {
-        showToUser(getIndexedListForViewing(list));
+    public static String formatIndexedList(List<String> list) {
+        return formatIndexedListForViewing(list);
     }
     /** Formats a list of strings as a viewable indexed list. */
-    private static String getIndexedListForViewing(List<String> listItems) {
+    private static String formatIndexedListForViewing(List<String> listItems) {
         final StringBuilder formatted = new StringBuilder();
         int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
         for (String listItem : listItems) {
-            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+            formatted.append(formatIndexedListItem(displayIndex, listItem)).append("\n");
             displayIndex++;
         }
         return formatted.toString();
@@ -101,7 +53,7 @@ public class Formatter {
      *
      * @param visibleIndex visible index for this listing
      */
-    private static String getIndexedListItem(int visibleIndex, String listItem) {
+    private static String formatIndexedListItem(int visibleIndex, String listItem) {
         return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
     }
