+++ src/seedu/addressbook/ui/TextUi.java
@@ -6,6 +6,12 @@ import static seedu.addressbook.common.Messages.MESSAGE_PROGRAM_LAUNCH_ARGS_USAG
 import static seedu.addressbook.common.Messages.MESSAGE_USING_STORAGE_FILE;
 import static seedu.addressbook.common.Messages.MESSAGE_WELCOME;
 
+import static seedu.addressbook.ui.Formatter.formatIndexedList;
+import static seedu.addressbook.ui.Formatter.formatPersonListView;
+import static seedu.addressbook.ui.Formatter.DIVIDER;
+import static seedu.addressbook.ui.Formatter.LINE_PREFIX;
+import static seedu.addressbook.ui.Formatter.LS;
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.ArrayList;
@@ -15,7 +21,6 @@ import java.util.Scanner;
 
 import seedu.addressbook.commands.CommandResult;
 import seedu.addressbook.data.person.ReadOnlyPerson;
-import seedu.addressbook.ui.Formatter;
 
 /**
  * Text UI of the application.
@@ -62,16 +67,67 @@ public class TextUi {
      * @return command (full line) entered by the user
      */
     public String getUserCommand() {
-        out.print(Formatter.LINE_PREFIX + "Enter command: ");
+        out.print(LINE_PREFIX + "Enter command: ");
         String fullInputLine = in.nextLine();
 
         // silently consume all ignored lines
         while (shouldIgnore(fullInputLine)) {
             fullInputLine = in.nextLine();
         }
-        Formatter formatter = new Formatter(out);
-        formatter.showToUser("[Command entered:" + fullInputLine + "]");
+        showToUser("[Command entered:" + fullInputLine + "]");
         return fullInputLine;
     }
+
+    public void showWelcomeMessage(String version, String storageFilePath) {
+        String storageFileInfo = String.format(MESSAGE_USING_STORAGE_FILE, storageFilePath);
+        showToUser(
+                DIVIDER,
+                DIVIDER,
+                MESSAGE_WELCOME,
+                version,
+                MESSAGE_PROGRAM_LAUNCH_ARGS_USAGE,
+                storageFileInfo,
+                DIVIDER);
+    }
+    public void showGoodbyeMessage() {
+        showToUser(MESSAGE_GOODBYE, DIVIDER, DIVIDER);
+    }
+
+    public void showInitFailedMessage() {
+        showToUser(MESSAGE_INIT_FAILED, DIVIDER, DIVIDER);
+    }
+
+    /** Shows message(s) to the user */
+    public void showToUser(String... message) {
+        for (String m : message) {
+            out.println(LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX));
+        }
+    }
+
+    /**
+     * Shows the result of a command execution to the user. Includes additional formatting to demarcate different
+     * command execution segments.
+     */
+    public void showResultToUser(CommandResult result) {
+        final Optional<List<? extends ReadOnlyPerson>> resultPersons = result.getRelevantPersons();
+        if (resultPersons.isPresent()) {
+            showToUser(formatPersonListView(resultPersons.get()));
+        }
+        showToUser(result.feedbackToUser, DIVIDER);
+    }
+
+    /**
+     * Shows a list of persons to the user, formatted as an indexed list.
+     * Private contact details are hidden.
+     */
+    private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
+        showToUser(formatPersonListView(persons));
+    }
+
+    /** Shows a list of strings to the user, formatted as an indexed list. */
+    private void showToUserAsIndexedList(List<String> list) {
+        showToUser(formatIndexedList(list));
+    }
+
 
 }
